Revert "sequencing_file"
